Create LISABIND
Founded by Daniel Martinez, LISABIND is a project that pretends to connect all kinds of local artists, (from Lisboa, Portugal). In order to alow this young artists to show there art skills and create the most unique nfts. As founder and artist I will be selling at first my artwork until i get more artists like me on board. NFTS will change the artistic world and I do believe that we artists are having low possibility to actually bring our art into nfts, as so Cardano is helping us a lot and with this project i pretend to brind as much artists to the cardano nft market as i can.